+++ alacritty_terminal/src/tty/mod.rs
@@ -50,9 +50,10 @@ impl Shell {
     }
 }
 
-/// This trait defines the behaviour needed to read and/or write to a stream.
-/// It defines an abstraction over polling's interface in order to allow either one
-/// read/write object or a separate read and write object.
+/// Stream read and/or write behavior.
+///
+/// This defines an abstraction over polling's interface in order to allow either
+/// one read/write object or a separate read and write object.
 pub trait EventedReadWrite {
     type Reader: io::Read;
     type Writer: io::Write;
@@ -97,10 +98,6 @@ pub fn setup_env() {
 
     // Advertise 24-bit color support.
     env::set_var("COLORTERM", "truecolor");
-
-    // Prevent child processes from inheriting startup notification env.
-    env::remove_var("DESKTOP_STARTUP_ID");
-    env::remove_var("XDG_ACTIVATION_TOKEN");
 }
 
 /// Check if a terminfo entry exists on the system.